feat(new-board): added a Presentation features which presents canvasboard cards as slides
Added a Presentation feature which utilizes Reveal.js library though which we create slides from
cards and do presentations utlizing them. The slides are added in a bootstrap modal which can be
then utlized to view and present the slides which are created dynamically everytime for now to
present. In future sync of slides and cards should be done using more lazy load approaches to avoid
unnessary compuatations. Press F for full screen and ESC for exiting the modal. More UI controls
should be added for these operations.